feat: 인기 게시글 컴포넌트 분할
