+++ Tlog/app/src/main/java/com/tlog/ui/screen/share/MainScreen.kt
@@ -19,7 +19,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.navigationBars
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
@@ -40,8 +39,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
@@ -58,16 +55,14 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import coil.compose.AsyncImage
 import com.tlog.R
-import com.tlog.data.model.share.Destination
 import com.tlog.data.model.share.LocationData
 import com.tlog.data.model.share.Post
-import com.tlog.data.model.share.RecommendDestination
 import com.tlog.ui.component.main.RecommendDestinationSection
+import com.tlog.ui.component.main.RecommendPostSection
 import com.tlog.ui.component.share.BottomBar
 import com.tlog.ui.component.share.MainTopBar
 import com.tlog.ui.component.travel.BlueHashTagGroup
 import com.tlog.ui.navigation.Screen
-import com.tlog.ui.style.Body1Bold
 import com.tlog.ui.style.BodyTitle
 import com.tlog.ui.theme.Essential
 import com.tlog.ui.theme.MainFont
@@ -369,106 +364,106 @@ fun MainScreen(
 
 
 
-                Spacer(modifier = Modifier.height(height = 49.dp))
-
-
-                data class Course(
-                    val title: String,
-                    val description: String,
-                    val image: Int
-                )
-
-                val popularCourseList = listOf(
-                    Course(
-                        title = "서울 청계천",
-                        description = "서울 중심에 위치한 청계천 코스",
-                        image = R.drawable.destination_img
-                    ),
-                    Course(
-                        title = "서울",
-                        description = "서울 중심에 위치한 청계천 코스",
-                        image = R.drawable.destination_img
-                    ),
-                    Course(
-                        title = "광명 여행",
-                        description = "서울 중심에 위치한 청계천 코스",
-                        image = R.drawable.destination_img
-                    ),
-                    Course(
-                        title = "대구 이월드 부근",
-                        description = "서울 중심에 위치한 청계천 코스",
-                        image = R.drawable.destination_img
-                    )
-                )
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 24.dp)
-                ) {
-                    Column(
-                        verticalArrangement = Arrangement.spacedBy(space = 27.dp)
-                    ) {
-                        Text(
-                            text = "인기 코스",
-                            style = BodyTitle
-                        )
-
-                        LazyRow(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .height(height = 146.dp)
-                        ) {
-                            items(
-                                items = popularCourseList,
-                                key = { course -> course.title }
-                            ) { course ->
-                                Box(
-                                    modifier = Modifier
-                                        .padding(end = 16.dp)
-                                        .clip(shape = RoundedCornerShape(size = 10.dp))
-                                        .background(color = Color.Gray)
-                                        .clip(shape = RoundedCornerShape(size = 10.dp))
-                                        .clickable { }
-                                ) {
-                                    Image(
-                                        painter = painterResource(id = course.image),
-                                        contentDescription = null,
-                                        contentScale = ContentScale.Crop,
-                                        modifier = Modifier
-                                            .fillMaxSize()
-                                    )
-
-                                    Column(
-                                        verticalArrangement = Arrangement.spacedBy(space = 30.dp),
-                                        modifier = Modifier
-                                            .fillMaxSize()
-                                            .padding(top = 24.dp, start = 24.dp)
-                                    ) {
-                                        Text(
-                                            text = course.title,
-                                            style = TextStyle(
-                                                fontFamily = MainFont,
-                                                fontSize = 24.sp,
-                                                fontWeight = FontWeight.Bold,
-                                                color = Color.White
-                                            )
-                                        )
-
-                                        Text(
-                                            text = course.description,
-                                            style = TextStyle(
-                                                fontFamily = MainFont,
-                                                fontSize = 13.sp,
-                                                fontWeight = FontWeight.Light,
-                                                color = Color.White
-                                            )
-                                        )
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
+//                Spacer(modifier = Modifier.height(height = 49.dp))
+//
+//
+//                data class Course(
+//                    val title: String,
+//                    val description: String,
+//                    val image: Int
+//                )
+//
+//                val popularCourseList = listOf(
+//                    Course(
+//                        title = "서울 청계천",
+//                        description = "서울 중심에 위치한 청계천 코스",
+//                        image = R.drawable.destination_img
+//                    ),
+//                    Course(
+//                        title = "서울",
+//                        description = "서울 중심에 위치한 청계천 코스",
+//                        image = R.drawable.destination_img
+//                    ),
+//                    Course(
+//                        title = "광명 여행",
+//                        description = "서울 중심에 위치한 청계천 코스",
+//                        image = R.drawable.destination_img
+//                    ),
+//                    Course(
+//                        title = "대구 이월드 부근",
+//                        description = "서울 중심에 위치한 청계천 코스",
+//                        image = R.drawable.destination_img
+//                    )
+//                )
+//                Box(
+//                    modifier = Modifier
+//                        .fillMaxWidth()
+//                        .padding(horizontal = 24.dp)
+//                ) {
+//                    Column(
+//                        verticalArrangement = Arrangement.spacedBy(space = 27.dp)
+//                    ) {
+//                        Text(
+//                            text = "인기 코스",
+//                            style = BodyTitle
+//                        )
+//
+//                        LazyRow(
+//                            modifier = Modifier
+//                                .fillMaxWidth()
+//                                .height(height = 146.dp)
+//                        ) {
+//                            items(
+//                                items = popularCourseList,
+//                                key = { course -> course.title }
+//                            ) { course ->
+//                                Box(
+//                                    modifier = Modifier
+//                                        .padding(end = 16.dp)
+//                                        .clip(shape = RoundedCornerShape(size = 10.dp))
+//                                        .background(color = Color.Gray)
+//                                        .clip(shape = RoundedCornerShape(size = 10.dp))
+//                                        .clickable { }
+//                                ) {
+//                                    Image(
+//                                        painter = painterResource(id = course.image),
+//                                        contentDescription = null,
+//                                        contentScale = ContentScale.Crop,
+//                                        modifier = Modifier
+//                                            .fillMaxSize()
+//                                    )
+//
+//                                    Column(
+//                                        verticalArrangement = Arrangement.spacedBy(space = 30.dp),
+//                                        modifier = Modifier
+//                                            .fillMaxSize()
+//                                            .padding(top = 24.dp, start = 24.dp)
+//                                    ) {
+//                                        Text(
+//                                            text = course.title,
+//                                            style = TextStyle(
+//                                                fontFamily = MainFont,
+//                                                fontSize = 24.sp,
+//                                                fontWeight = FontWeight.Bold,
+//                                                color = Color.White
+//                                            )
+//                                        )
+//
+//                                        Text(
+//                                            text = course.description,
+//                                            style = TextStyle(
+//                                                fontFamily = MainFont,
+//                                                fontSize = 13.sp,
+//                                                fontWeight = FontWeight.Light,
+//                                                color = Color.White
+//                                            )
+//                                        )
+//                                    }
+//                                }
+//                            }
+//                        }
+//                    }
+//                }
 
 
 
@@ -491,61 +486,8 @@ fun MainScreen(
 
                 Spacer(modifier = Modifier.height(43.dp))
 
+                RecommendPostSection(recommendPosts)
 
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                ) {
-                    Text(
-                        text = "인기 게시글",
-                        style = BodyTitle,
-                        modifier = Modifier
-                            .padding(start = 24.dp)
-                    )
-
-                    Spacer(modifier = Modifier.height(28.dp))
-
-                    LazyRow(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 28.dp),
-                        horizontalArrangement = Arrangement.spacedBy(30.dp)
-                    ) {
-                        items(
-                            items = recommendPosts,
-                            key = { post -> post.title }
-                        ) { post ->
-                            RecommendPostCard(post)
-                        }
-                    }
-
-                    Spacer(modifier = Modifier.height(28.dp))
-
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(end = 16.dp),
-                        horizontalArrangement = Arrangement.End
-                    ) {
-                        Text(
-                            text = "게시글 더 보러가기",
-                            style = TextStyle(
-                                fontFamily = MainFont,
-                                fontSize = 13.sp,
-                                color = Essential,
-                                fontWeight = FontWeight.SemiBold
-                            )
-                        )
-
-                        Icon(
-                            painter = painterResource(id = R.drawable.ic_arrow_right),
-                            contentDescription = "right_arrow",
-                            tint = Essential,
-                            modifier = Modifier
-                                .size(20.dp)
-                        )
-                    }
-                }
 
 
                 // ISSUE
@@ -633,65 +575,3 @@ fun MainScreen(
     }
 }
 
-@Composable
-fun RecommendPostCard(
-    post: Post
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-    ) {
-        Row(
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-            modifier = Modifier
-                .fillMaxWidth()
-        ) {
-            AsyncImage(
-                model = post.imageUrls.firstOrNull() ?: "",
-                contentDescription = null,
-                contentScale = ContentScale.Crop,
-                error = painterResource(id = R.drawable.tmp_jeju),
-                modifier = Modifier
-                    .height(158.dp)
-                    .width(150.dp)
-                    .clip(RoundedCornerShape(10.dp))
-            )
-            if (post.imageUrls.size > 1) {
-                AsyncImage(
-                    model = post.imageUrls[1],
-                    contentDescription = null,
-                    contentScale = ContentScale.Crop,
-                    error = painterResource(id = R.drawable.tmp_jeju),
-                    modifier = Modifier
-                        .height(158.dp)
-                        .width(150.dp)
-                        .clip(RoundedCornerShape(10.dp))
-                )
-            }
-        }
-
-        Spacer(modifier = Modifier.height(14.dp))
-
-        Text(
-            text = post.title,
-            style = TextStyle(
-                fontFamily = MainFont,
-                fontSize = 16.sp,
-                fontWeight = FontWeight.Bold
-            )
-        )
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Text(
-            text = post.description,
-            style = TextStyle(
-                fontFamily = MainFont,
-                fontSize = 14.sp,
-                color = Color.Gray
-            ),
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis
-        )
-    }
-}
